+++ modules/ans_infer/Transformer.py
@@ -17,30 +17,30 @@ class TransDecoder(torch_nn.Module):
         self.max_len_ans= args.max_len_ans
         self.seq_len_ans= args.seq_len_ans
 
-        decoder_layer   = torch_nn.TransformerDecoderLayer(d_model=self.d_hid*2, nhead=8)
-        self.decoder    = torch_nn.TransformerDecoder(decoder_layer, num_layers=6)
+        decoder_layer   = torch_nn.TransformerDecoderLayer(d_model=self.d_hid, nhead=args.trans_nheads)
+        self.decoder    = torch_nn.TransformerDecoder(decoder_layer, num_layers=args.trans_nlayers)
 
         self.embedding  = bert_model
-        self.ff_ans     = torch_nn.Linear(200, self.d_hid*2, bias=False)
+        self.ff_ans     = torch_nn.Linear(self.d_hid, self.d_hid, bias=False)
 
         self.ff_pred    = torch_nn.Sequential(
-            torch_nn.Linear(self.d_hid*2, self.d_vocab),
-            torch_nn.Tanh(),
+            torch_nn.Linear(self.d_hid, self.d_vocab),
+            torch_nn.GELU(),
             torch_nn.Dropout(args.dropout)
         )
 
     def forward(self, Y, ans, is_inferring=False):
-        # Y         : [b, seq_len_contx, d_hid * 2]
-        # ans       : [b, seq_len_ans, 200]
+        # Y         : [n_paras+1, b, d_hid]
+        # ans       : [b, seq_len_ans, d_hid]
 
         batch   = Y.shape[0]
 
         if is_inferring:
             def infer(tok_ids, Y):
-                # Y      : [seq_len_contx, d_hid * 2]
+                # Y      : [n_paras+1, d_hid]
                 # tok_ids: list of token ids
-                Y      = Y.unsqueeze(0).transpose(0, 1)
-                # [seq_len_contx, b=1, d_hid*2]
+                Y      = Y.unsqueeze(1)
+                # [n_paras+1, b=1, d_hid]
 
                 toks_emb = torch.LongTensor(tok_ids).unsqueeze(0).to(args.device)
                 toks_emb = self.embedding(toks_emb)[0]
@@ -49,13 +49,13 @@ class TransDecoder(torch_nn.Module):
                 # [b=1, seq=*, 768]
 
                 toks_emb = self.ff_ans(toks_emb).transpose(0, 1)
-                # [seq=*, b=1, d_hid*2]
+                # [seq=*, b=1, d_hid]
 
                 output  = self.decoder(toks_emb, Y)
-                # [seq=*, b=1, d_hid*2]
+                # [seq=*, b=1, d_hid]
 
                 output  = output.transpose(0, 1)
-                # [b=1, seq=*, d_hid*2]
+                # [b=1, seq=*, d_hid]
 
                 output  = self.ff_pred(output).squeeze(0)
                 # [seq=*, d_vocab]
@@ -67,23 +67,20 @@ class TransDecoder(torch_nn.Module):
                                      topk_strategy="select_mix_beam")
 
             for b in range(batch):
-                indices = beam_search.search(Y[b, :, :])
+                indices = beam_search.search(Y[:, b, :])
 
                 pred.append(indices)
 
             return pred
 
         else:
-            Y       = Y.transpose(0, 1)
-            # [seq_len_contx, b, d_hid*2]
-
             ans     = self.ff_ans(ans).transpose(0, 1)
-            # [seq_len_ans, b, d_hid*2]
+            # [seq_len_ans, b, d_hid]
 
             pred    = self.decoder(ans, Y)
-            # [seq_len_ans, b, d_hid*2]
+            # [seq_len_ans, b, d_hid]
             pred    = pred.transpose(0, 1)
-            # [b, seq_len_ans, d_hid*2]
+            # [b, seq_len_ans, d_hid]
 
             pred    = self.ff_pred(pred)
             # [b, seq_len_ans, d_vocab]