+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ### This file contains changes and modifications during developing.
 
+## May 15 2021, 10:26
+- Take *GloVe embedding* back and modify class *Vocab*
+- Edit some fields during training
+- Enhance some args in 'configs.py'
+- Add Embedding-enhanced module *BertBasedEmbd*
+- Modify output in module *MemoryBased*: ouput now is not just from memory but combination of it with outputs from each para.
+
+## May 14 2021, 10:26
+- Add new reasoning module: *MemoryBased* that is based on Memory network
+- Modify in module *FineGrain*: convert from NN to Linear
+- Modify module *TransDecoder* and *NarrativePipeline* to fit with new reasoning module
+
 ## May 11 2021, 16:32
 - Add new implementation of *BeamSearch*
 - Simplify `launcher.sh`